implement abstract base class for event sub-process start event behaviors to make sonarqube happy
