+++ activiti-core/activiti-engine/src/main/java/org/activiti/engine/impl/bpmn/behavior/EventSubProcessMessageStartEventActivityBehavior.java
@@ -15,33 +15,17 @@
  */
 package org.activiti.engine.impl.bpmn.behavior;
 
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import org.activiti.bpmn.model.EventSubProcess;
 import org.activiti.bpmn.model.MessageEventDefinition;
-import org.activiti.bpmn.model.StartEvent;
-import org.activiti.bpmn.model.SubProcess;
-import org.activiti.bpmn.model.ValuedDataObject;
 import org.activiti.engine.delegate.DelegateExecution;
-import org.activiti.engine.delegate.event.impl.ActivitiEventBuilder;
-import org.activiti.engine.history.DeleteReason;
 import org.activiti.engine.impl.bpmn.parser.factory.MessageExecutionContext;
-import org.activiti.engine.impl.context.Context;
-import org.activiti.engine.impl.interceptor.CommandContext;
 import org.activiti.engine.impl.persistence.entity.EventSubscriptionEntity;
-import org.activiti.engine.impl.persistence.entity.EventSubscriptionEntityManager;
-import org.activiti.engine.impl.persistence.entity.ExecutionEntity;
-import org.activiti.engine.impl.persistence.entity.ExecutionEntityManager;
 import org.activiti.engine.impl.persistence.entity.MessageEventSubscriptionEntity;
 
 /**
- * Implementation of the BPMN 2.0 event subprocess message start event.
- *
-
+ * Implementation of the BPMN 2.0 event sub-process message start event.
  */
-public class EventSubProcessMessageStartEventActivityBehavior extends AbstractBpmnActivityBehavior {
+public class EventSubProcessMessageStartEventActivityBehavior
+    extends AbstractEventSubProcessStartEventActivityBehavior {
 
     private static final long serialVersionUID = 1L;
 
@@ -56,74 +40,17 @@ public class EventSubProcessMessageStartEventActivityBehavior extends AbstractBp
         this.messageExecutionContext = messageExecutionContext;
     }
 
-    public void execute(DelegateExecution execution) {
-        StartEvent startEvent = (StartEvent) execution.getCurrentFlowElement();
-        EventSubProcess eventSubProcess = (EventSubProcess) startEvent.getSubProcess();
-
-        execution.setScope(true);
-
-        // initialize the template-defined data objects as variables
-        Map<String, Object> dataObjectVars = processDataObjects(eventSubProcess.getDataObjects());
-        if (dataObjectVars != null) {
-            execution.setVariablesLocal(dataObjectVars);
-        }
+    @Override
+    protected String resolveEventName(DelegateExecution execution) {
+        // Should we use triggerName and triggerData, because message name expression can change?
+        return messageExecutionContext.getMessageName(execution);
     }
 
     @Override
-    public void trigger(DelegateExecution execution, String triggerName, Object triggerData) {
-        CommandContext commandContext = Context.getCommandContext();
-        ExecutionEntityManager executionEntityManager = commandContext.getExecutionEntityManager();
-        ExecutionEntity executionEntity = (ExecutionEntity) execution;
-
-        StartEvent startEvent = (StartEvent) execution.getCurrentFlowElement();
-        if (startEvent.isInterrupting()) {
-            List<ExecutionEntity> childExecutions = executionEntityManager.findChildExecutionsByParentExecutionId(
-                executionEntity.getParentId()
-            );
-            for (ExecutionEntity childExecution : childExecutions) {
-                if (!childExecution.getId().equals(executionEntity.getId())) {
-                    executionEntityManager.cancelExecutionAndRelatedData(
-                        childExecution,
-                        DeleteReason.EVENT_SUBPROCESS_INTERRUPTING + "(" + startEvent.getId() + ")"
-                    );
-                }
-            }
-        }
-
-        // Should we use triggerName and triggerData, because message name expression can change?
-        String messageName = messageExecutionContext.getMessageName(execution);
-
-        EventSubscriptionEntityManager eventSubscriptionEntityManager =
-            Context.getCommandContext().getEventSubscriptionEntityManager();
-        List<EventSubscriptionEntity> eventSubscriptions = executionEntity.getEventSubscriptions();
-        for (EventSubscriptionEntity eventSubscription : eventSubscriptions) {
-            if (
-                eventSubscription instanceof MessageEventSubscriptionEntity &&
-                eventSubscription.getEventName().equals(messageName)
-            ) {
-                eventSubscriptionEntityManager.delete(eventSubscription);
-            }
-        }
-
-        executionEntity.setCurrentFlowElement(
-            (SubProcess) executionEntity.getCurrentFlowElement().getParentContainer()
+    protected boolean matchesSubscription(EventSubscriptionEntity eventSubscription, String eventName) {
+        return (
+            eventSubscription instanceof MessageEventSubscriptionEntity &&
+            eventSubscription.getEventName().equals(eventName)
         );
-        executionEntity.setScope(true);
-
-        ExecutionEntity outgoingFlowExecution = executionEntityManager.createChildExecution(executionEntity);
-        outgoingFlowExecution.setCurrentFlowElement(startEvent);
-
-        leave(outgoingFlowExecution);
-    }
-
-    protected Map<String, Object> processDataObjects(Collection<ValuedDataObject> dataObjects) {
-        Map<String, Object> variablesMap = new HashMap<>();
-        // convert data objects to process variables
-        if (dataObjects != null) {
-            for (ValuedDataObject dataObject : dataObjects) {
-                variablesMap.put(dataObject.getName(), dataObject.getValue());
-            }
-        }
-        return variablesMap;
     }
 }
