+++ activiti-core/activiti-engine/src/main/java/org/activiti/engine/impl/bpmn/behavior/EventSubProcessSignalStartEventActivityBehavior.java
@@ -15,20 +15,10 @@
  */
 package org.activiti.engine.impl.bpmn.behavior;
 
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import org.activiti.bpmn.model.EventSubProcess;
 import org.activiti.bpmn.model.Signal;
 import org.activiti.bpmn.model.SignalEventDefinition;
 import org.activiti.bpmn.model.StartEvent;
-import org.activiti.bpmn.model.SubProcess;
-import org.activiti.bpmn.model.ValuedDataObject;
 import org.activiti.engine.delegate.DelegateExecution;
-import org.activiti.engine.history.DeleteReason;
-import org.activiti.engine.impl.context.Context;
-import org.activiti.engine.impl.interceptor.CommandContext;
 import org.activiti.engine.impl.persistence.entity.EventSubscriptionEntity;
 import org.activiti.engine.impl.persistence.entity.EventSubscriptionEntityManager;
 import org.activiti.engine.impl.persistence.entity.ExecutionEntity;
@@ -39,7 +29,7 @@ import org.activiti.engine.impl.persistence.entity.SignalEventSubscriptionEntity
  * Implementation of the BPMN 2.0 event sub-process signal start event.
  * Mirrors {@link EventSubProcessMessageStartEventActivityBehavior} but for {@link SignalEventDefinition}.
  */
-public class EventSubProcessSignalStartEventActivityBehavior extends AbstractBpmnActivityBehavior {
+public class EventSubProcessSignalStartEventActivityBehavior extends AbstractEventSubProcessStartEventActivityBehavior {
 
     private static final long serialVersionUID = 1L;
 
@@ -52,84 +42,37 @@ public class EventSubProcessSignalStartEventActivityBehavior extends AbstractBpm
     }
 
     @Override
-    public void execute(DelegateExecution execution) {
-        StartEvent startEvent = (StartEvent) execution.getCurrentFlowElement();
-        EventSubProcess eventSubProcess = (EventSubProcess) startEvent.getSubProcess();
-
-        execution.setScope(true);
-
-        // initialize the template-defined data objects as variables
-        Map<String, Object> dataObjectVars = processDataObjects(eventSubProcess.getDataObjects());
-        if (dataObjectVars != null) {
-            execution.setVariablesLocal(dataObjectVars);
-        }
+    protected String resolveEventName(DelegateExecution execution) {
+        return (signal != null) ? signal.getName() : signalEventDefinition.getSignalRef();
     }
 
     @Override
-    public void trigger(DelegateExecution execution, String triggerName, Object triggerData) {
-        CommandContext commandContext = Context.getCommandContext();
-        ExecutionEntityManager executionEntityManager = commandContext.getExecutionEntityManager();
-        ExecutionEntity executionEntity = (ExecutionEntity) execution;
-
-        StartEvent startEvent = (StartEvent) execution.getCurrentFlowElement();
-        if (startEvent.isInterrupting()) {
-            List<ExecutionEntity> childExecutions = executionEntityManager.findChildExecutionsByParentExecutionId(
-                executionEntity.getParentId()
-            );
-            for (ExecutionEntity childExecution : childExecutions) {
-                if (!childExecution.getId().equals(executionEntity.getId())) {
-                    executionEntityManager.cancelExecutionAndRelatedData(
-                        childExecution,
-                        DeleteReason.EVENT_SUBPROCESS_INTERRUPTING + "(" + startEvent.getId() + ")"
-                    );
-                }
-            }
-        }
-
-        String signalName = (signal != null) ? signal.getName() : signalEventDefinition.getSignalRef();
-
-        EventSubscriptionEntityManager eventSubscriptionEntityManager =
-            commandContext.getEventSubscriptionEntityManager();
-        List<EventSubscriptionEntity> eventSubscriptions = executionEntity.getEventSubscriptions();
-        for (EventSubscriptionEntity eventSubscription : eventSubscriptions) {
-            if (
-                eventSubscription instanceof SignalEventSubscriptionEntity &&
-                eventSubscription.getEventName().equals(signalName)
-            ) {
-                // Interrupting: the event scope is gone after we leave; drop the subscription.
-                // Non-interrupting: we need to keep listening for further signals, so re-create
-                // an equivalent subscription on the parent process instance after this one is consumed.
-                eventSubscriptionEntityManager.delete(eventSubscription);
-                if (!startEvent.isInterrupting()) {
-                    ExecutionEntity parent = executionEntity.getParent();
-                    if (parent != null) {
-                        ExecutionEntity newEventScope = executionEntityManager.createChildExecution(parent);
-                        newEventScope.setCurrentFlowElement(startEvent);
-                        newEventScope.setEventScope(true);
-                        eventSubscriptionEntityManager.insertSignalEvent(signalName, signal, newEventScope);
-                    }
-                }
-            }
-        }
-
-        executionEntity.setCurrentFlowElement(
-            (SubProcess) executionEntity.getCurrentFlowElement().getParentContainer()
+    protected boolean matchesSubscription(EventSubscriptionEntity eventSubscription, String eventName) {
+        return (
+            eventSubscription instanceof SignalEventSubscriptionEntity &&
+            eventSubscription.getEventName().equals(eventName)
         );
-        executionEntity.setScope(true);
-
-        ExecutionEntity outgoingFlowExecution = executionEntityManager.createChildExecution(executionEntity);
-        outgoingFlowExecution.setCurrentFlowElement(startEvent);
-
-        leave(outgoingFlowExecution);
     }
 
-    protected Map<String, Object> processDataObjects(Collection<ValuedDataObject> dataObjects) {
-        Map<String, Object> variablesMap = new HashMap<>();
-        if (dataObjects != null) {
-            for (ValuedDataObject dataObject : dataObjects) {
-                variablesMap.put(dataObject.getName(), dataObject.getValue());
+    @Override
+    protected void onSubscriptionConsumed(
+        ExecutionEntity executionEntity,
+        StartEvent startEvent,
+        String eventName,
+        ExecutionEntityManager executionEntityManager,
+        EventSubscriptionEntityManager eventSubscriptionEntityManager
+    ) {
+        // Interrupting: the event scope is gone after we leave; nothing more to do.
+        // Non-interrupting: keep listening for further signals by re-arming an equivalent
+        // subscription on the parent process instance after this one is consumed.
+        if (!startEvent.isInterrupting()) {
+            ExecutionEntity parent = executionEntity.getParent();
+            if (parent != null) {
+                ExecutionEntity newEventScope = executionEntityManager.createChildExecution(parent);
+                newEventScope.setCurrentFlowElement(startEvent);
+                newEventScope.setEventScope(true);
+                eventSubscriptionEntityManager.insertSignalEvent(eventName, signal, newEventScope);
             }
         }
-        return variablesMap;
     }
 }
